minor fixes

+++ backend/src/controllers/transaction.js
@@ -1,10 +1,11 @@
 const Transaction = require('../models/Transaction')
-
-const cloudinary = require('cloudinary').v2
+const { uploadImage, deleteImage } = require('../utils/image')
+const { returnReadableDate } = require('../utils/date')
 
 module.exports = {
   async index(req, res) {
     const transactions = await Transaction.find()
+
     if (transactions === []) {
       return res.status(404).send('Nenhuma transação encontrada')
     }
@@ -20,17 +21,14 @@ module.exports = {
         .send('Erro de Validação: Preencha os campos necessários')
     } else {
       const dateString = await returnReadableDate(date)
-      let imageData = {
-        public_id: null,
-        imageSrc: undefined
-      }
+
+      let imageData = {}
       if (image) {
-        imageData = await uploadToCloudinary(image)
+        imageData = await uploadImage(image)
       }
 
       const response = await Transaction.create({
         ...imageData,
-
         type,
         member,
         value,
@@ -42,42 +40,30 @@ module.exports = {
     }
   },
   async edit(req, res) {
-    let {
-      value,
-      type,
-      member,
-      description,
-      imageSrc,
-      date,
-      public_id,
-      newImage
-    } = req.body
+    let { imageSrc, date, public_id, newImage } = req.body
 
     const old = await Transaction.findById(req.params._id)
+
     if (old.date !== date) {
       date = await returnReadableDate(date)
     }
+
     let imageData = {
-      public_id: null,
-      imageSrc: undefined
-    }
-    if (newImage) {
-      imageData = await changeCloudinaryImage(newImage, public_id)
+      public_id,
+      imageSrc
     }
 
-    public_id = imageData.public_id
-    imageSrc = imageData.imageSrc
+    if (newImage) {
+      await deleteImage(public_id)
+      imageData = await uploadImage(newImage)
+    }
 
     const newTransaction = await Transaction.findOneAndUpdate(
       { _id: req.params._id },
       {
-        value,
-        type,
-        date,
-        member,
-        description,
-        imageSrc,
-        public_id
+        ...req.body,
+        ...imageData,
+        date
       },
       { new: true }
     )
@@ -86,61 +72,14 @@ module.exports = {
   },
   async delete(req, res) {
     const _id = req.params._id
-
-    const response = await Transaction.findByIdAndDelete(_id)
-    if (response.public_id) {
-      await cloudinary.uploader.destroy(response.public_id)
-    }
-    return res.send(response._id)
-  }
-}
-
-const returnReadableDate = async date => {
-  const parsedDate = new Date(date)
-  const dateString = parsedDate.toLocaleDateString('pt-BR', {
-    dateStyle: 'short'
-  })
-  const timeString = parsedDate.toLocaleTimeString('pt-BR', {
-    timeStyle: 'short'
-  })
-  const nums = dateString.split('-').reverse().join('/')
-
-  return `${nums} ${timeString}`
-}
-
-const uploadToCloudinary = async base64Img => {
-  try {
-    const { public_id, url } = await cloudinary.uploader.upload(base64Img, {})
-
-    return {
-      public_id,
-      imageSrc: url
-    }
-  } catch (error) {
-    console.log('error on cloudinary upload')
-    return {
-      public_id: '',
-      imageSrc: ''
-    }
-  }
-}
-
-const changeCloudinaryImage = async (base64Img, previousImg_id = null) => {
-  try {
-    if (previousImg_id) {
-      await cloudinary.uploader.destroy(previousImg_id)
-    }
-    const { public_id, url } = await cloudinary.uploader.upload(base64Img, {})
-
-    return {
-      public_id,
-      imageSrc: url
-    }
-  } catch (error) {
-    console.log('error on cloudinary upload')
-    return {
-      public_id: '',
-      imageSrc: ''
+    if (!_id) {
+      return res.status(400).send('Erro ao deletar: identificador vazio')
+    } else {
+      const response = await Transaction.findByIdAndDelete(_id)
+      if (response.public_id) {
+        await deleteImage(response.public_id)
+      }
+      return res.send(response._id)
     }
   }
 }
